fix keeper tests
