+++ x/crosschain/keeper/gas_payment_test.go
@@ -7,12 +7,7 @@ import (
 	"cosmossdk.io/math"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
-	"github.com/ethereum/go-ethereum/common"
-	evmkeeper "github.com/evmos/ethermint/x/evm/keeper"
 	"github.com/stretchr/testify/require"
-	"github.com/zeta-chain/protocol-contracts/pkg/uniswap/v2-periphery/contracts/uniswapv2router02.sol"
-	"github.com/zeta-chain/zetacore/cmd/zetacored/config"
 	zetacommon "github.com/zeta-chain/zetacore/common"
 	testkeeper "github.com/zeta-chain/zetacore/testutil/keeper"
 	"github.com/zeta-chain/zetacore/testutil/sample"
@@ -21,193 +16,6 @@ import (
 	fungibletypes "github.com/zeta-chain/zetacore/x/fungible/types"
 	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 )
-
-// get a valid eth chain id independently of the build flag
-func getValidEthChainID(t *testing.T) int64 {
-	return getValidEthChain(t).ChainId
-}
-
-// get a valid eth chain independently of the build flag
-func getValidEthChain(_ *testing.T) *zetacommon.Chain {
-	goerli := zetacommon.GoerliLocalnetChain()
-	return &goerli
-}
-
-// assert that a contract has been deployed by checking stored code is non-empty.
-func assertContractDeployment(t *testing.T, k *evmkeeper.Keeper, ctx sdk.Context, contractAddress common.Address) {
-	acc := k.GetAccount(ctx, contractAddress)
-	require.NotNil(t, acc)
-
-	code := k.GetCode(ctx, common.BytesToHash(acc.CodeHash))
-	require.NotEmpty(t, code)
-}
-
-// deploySystemContracts deploys the system contracts and returns their addresses.
-func deploySystemContracts(
-	t *testing.T,
-	ctx sdk.Context,
-	k *fungiblekeeper.Keeper,
-	evmk *evmkeeper.Keeper,
-) (wzeta, uniswapV2Factory, uniswapV2Router, connector, systemContract common.Address) {
-	var err error
-
-	wzeta, err = k.DeployWZETA(ctx)
-	require.NoError(t, err)
-	require.NotEmpty(t, wzeta)
-	assertContractDeployment(t, evmk, ctx, wzeta)
-
-	uniswapV2Factory, err = k.DeployUniswapV2Factory(ctx)
-	require.NoError(t, err)
-	require.NotEmpty(t, uniswapV2Factory)
-	assertContractDeployment(t, evmk, ctx, uniswapV2Factory)
-
-	uniswapV2Router, err = k.DeployUniswapV2Router02(ctx, uniswapV2Factory, wzeta)
-	require.NoError(t, err)
-	require.NotEmpty(t, uniswapV2Router)
-	assertContractDeployment(t, evmk, ctx, uniswapV2Router)
-
-	connector, err = k.DeployConnectorZEVM(ctx, wzeta)
-	require.NoError(t, err)
-	require.NotEmpty(t, connector)
-	assertContractDeployment(t, evmk, ctx, connector)
-
-	systemContract, err = k.DeploySystemContract(ctx, wzeta, uniswapV2Factory, uniswapV2Router)
-	require.NoError(t, err)
-	require.NotEmpty(t, systemContract)
-	assertContractDeployment(t, evmk, ctx, systemContract)
-
-	return
-}
-
-// setupGasCoin is a helper function to setup the gas coin for testing
-func setupGasCoin(
-	t *testing.T,
-	ctx sdk.Context,
-	k *fungiblekeeper.Keeper,
-	evmk *evmkeeper.Keeper,
-	chainID int64,
-	assetName string,
-	symbol string,
-) (zrc20 common.Address) {
-	addr, err := k.SetupChainGasCoinAndPool(
-		ctx,
-		chainID,
-		assetName,
-		symbol,
-		8,
-		nil,
-	)
-	require.NoError(t, err)
-	assertContractDeployment(t, evmk, ctx, addr)
-	return addr
-}
-
-// deployZRC20 deploys a ZRC20 contract and returns its address
-func deployZRC20(
-	t *testing.T,
-	ctx sdk.Context,
-	k *fungiblekeeper.Keeper,
-	evmk *evmkeeper.Keeper,
-	chainID int64,
-	assetName string,
-	assetAddress string,
-	symbol string,
-) (zrc20 common.Address) {
-	addr, err := k.DeployZRC20Contract(
-		ctx,
-		assetName,
-		symbol,
-		8,
-		chainID,
-		0,
-		assetAddress,
-		big.NewInt(21_000),
-	)
-	require.NoError(t, err)
-	assertContractDeployment(t, evmk, ctx, addr)
-	return addr
-}
-
-// setupZRC20Pool setup a Uniswap pool with liquidity for the pair zeta/asset
-func setupZRC20Pool(
-	t *testing.T,
-	ctx sdk.Context,
-	k *fungiblekeeper.Keeper,
-	bankKeeper bankkeeper.Keeper,
-	zrc20Addr common.Address,
-) {
-	routerAddress, err := k.GetUniswapV2Router02Address(ctx)
-	require.NoError(t, err)
-	routerABI, err := uniswapv2router02.UniswapV2Router02MetaData.GetAbi()
-	require.NoError(t, err)
-
-	// enough for the small numbers used in test
-	liquidityAmount := big.NewInt(1e17)
-
-	// mint some zrc20 and zeta
-	_, err = k.DepositZRC20(ctx, zrc20Addr, types.ModuleAddressEVM, liquidityAmount)
-	require.NoError(t, err)
-	err = bankKeeper.MintCoins(
-		ctx,
-		types.ModuleName,
-		sdk.NewCoins(sdk.NewCoin(config.BaseDenom, sdk.NewIntFromBigInt(liquidityAmount))),
-	)
-	require.NoError(t, err)
-
-	// approve the router to spend the zeta
-	err = k.CallZRC20Approve(
-		ctx,
-		types.ModuleAddressEVM,
-		zrc20Addr,
-		routerAddress,
-		liquidityAmount,
-		false,
-	)
-	require.NoError(t, err)
-
-	// add the liquidity
-	//function addLiquidityETH(
-	//	address token,
-	//	uint amountTokenDesired,
-	//	uint amountTokenMin,
-	//	uint amountETHMin,
-	//	address to,
-	//	uint deadline
-	//)
-	_, err = k.CallEVM(
-		ctx,
-		*routerABI,
-		types.ModuleAddressEVM,
-		routerAddress,
-		liquidityAmount,
-		big.NewInt(5_000_000),
-		true,
-		false,
-		"addLiquidityETH",
-		zrc20Addr,
-		liquidityAmount,
-		fungiblekeeper.BigIntZero,
-		fungiblekeeper.BigIntZero,
-		types.ModuleAddressEVM,
-		liquidityAmount,
-	)
-	require.NoError(t, err)
-}
-
-func setAdminPolicies(ctx sdk.Context, zk testkeeper.ZetaKeepers, admin string) {
-	params := zk.ObserverKeeper.GetParams(ctx)
-	params.AdminPolicy = []*observertypes.Admin_Policy{
-		{
-			PolicyType: observertypes.Policy_Type_group1,
-			Address:    admin,
-		},
-		{
-			PolicyType: observertypes.Policy_Type_group2,
-			Address:    admin,
-		},
-	}
-	zk.ObserverKeeper.SetParams(ctx, params)
-}
 
 var (
 	// gasLimit = big.NewInt(21_000) - value used in SetupChainGasCoinAndPool for gas limit initialization
@@ -226,6 +34,8 @@ func TestKeeper_PayGasNativeAndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
+
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		zrc20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foobar", "foobar")
 		_, err := fungibleMsgServer.UpdateZRC20WithdrawFee(
@@ -294,6 +104,7 @@ func TestKeeper_PayGasNativeAndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foobar", "foobar")
 
@@ -325,6 +136,7 @@ func TestKeeper_PayGasNativeAndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		zrc20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foobar", "foobar")
 		_, err := fungibleMsgServer.UpdateZRC20WithdrawFee(
@@ -369,6 +181,7 @@ func TestKeeper_PayGasInERC20AndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin, erc20 and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		assetAddress := sample.EthAddress().String()
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		gasZRC20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foo", "foo")
@@ -462,6 +275,7 @@ func TestKeeper_PayGasInERC20AndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foobar", "foobar")
 
@@ -493,6 +307,7 @@ func TestKeeper_PayGasInERC20AndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin, erc20 and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		assetAddress := sample.EthAddress().String()
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		gasZRC20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foo", "foo")
@@ -538,6 +353,7 @@ func TestKeeper_PayGasInERC20AndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin, erc20 and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		assetAddress := sample.EthAddress().String()
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		gasZRC20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foo", "foo")
@@ -593,6 +409,7 @@ func TestKeeper_PayGasInERC20AndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin, erc20 and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		assetAddress := sample.EthAddress().String()
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		gasZRC20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foo", "foo")
@@ -663,6 +480,7 @@ func TestKeeper_PayGasInZetaAndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		zrc20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foobar", "foobar")
 		k.SetGasPrice(ctx, types.GasPrice{
@@ -751,6 +569,7 @@ func TestKeeper_PayGasInZetaAndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foobar", "foobar")
 
@@ -781,6 +600,7 @@ func TestKeeper_PayGasInZetaAndUpdateCctx(t *testing.T) {
 
 		// deploy gas coin and set fee params
 		chainID := getValidEthChainID(t)
+		setSupportedChain(ctx, zk, chainID)
 		deploySystemContracts(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper)
 		zrc20 := setupGasCoin(t, ctx, zk.FungibleKeeper, sdkk.EvmKeeper, chainID, "foobar", "foobar")
 		k.SetGasPrice(ctx, types.GasPrice{
