+++ x/crosschain/keeper/msg_server_update_tss_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"github.com/stretchr/testify/require"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -44,6 +45,7 @@ func TestMsgServer_UpdateTssAddress(t *testing.T) {
 		migrators := k.GetObserverKeeper().GetAllTssFundMigrators(ctx)
 		assert.Equal(t, 0, len(migrators))
 	})
+
 	t.Run("new tss has not been added to tss history", func(t *testing.T) {
 		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 		admin := sample.AccAddress()
@@ -75,6 +77,7 @@ func TestMsgServer_UpdateTssAddress(t *testing.T) {
 		assert.Equal(t, tssOld, tss)
 		assert.Equal(t, len(k.GetObserverKeeper().GetAllTssFundMigrators(ctx)), len(k.GetObserverKeeper().GetSupportedChains(ctx)))
 	})
+
 	t.Run("old tss pubkey provided", func(t *testing.T) {
 		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 		admin := sample.AccAddress()
@@ -105,6 +108,7 @@ func TestMsgServer_UpdateTssAddress(t *testing.T) {
 		assert.Equal(t, tssOld, tss)
 		assert.Equal(t, len(k.GetObserverKeeper().GetAllTssFundMigrators(ctx)), len(k.GetObserverKeeper().GetSupportedChains(ctx)))
 	})
+
 	t.Run("unable to update tss when not enough migrators are present", func(t *testing.T) {
 		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 		admin := sample.AccAddress()
@@ -112,33 +116,37 @@ func TestMsgServer_UpdateTssAddress(t *testing.T) {
 		msgServer := keeper.NewMsgServerImpl(*k)
 		tssOld := sample.Tss()
 		tssNew := sample.Tss()
+
 		k.GetObserverKeeper().SetTSSHistory(ctx, tssOld)
 		k.GetObserverKeeper().SetTSSHistory(ctx, tssNew)
 		k.GetObserverKeeper().SetTSS(ctx, tssOld)
-		for _, chain := range k.GetObserverKeeper().GetSupportedChains(ctx) {
-			index := chain.ChainName.String() + "_migration_tx_index"
-			k.GetObserverKeeper().SetFundMigrator(ctx, types.TssFundMigratorInfo{
-				ChainId:            chain.ChainId,
-				MigrationCctxIndex: index,
-			})
-			cctx := sample.CrossChainTx(t, index)
-			cctx.CctxStatus.Status = crosschaintypes.CctxStatus_OutboundMined
-			k.SetCrossChainTx(ctx, *cctx)
-			break
-		}
-		assert.Equal(t, len(k.GetObserverKeeper().GetAllTssFundMigrators(ctx)), 1)
+		setSupportedChain(ctx, zk, getValidEthChainIDWithIndex(t, 0), getValidEthChainIDWithIndex(t, 1))
+
+		// set a single migrator while there are 2 supported chains
+		chain := k.GetObserverKeeper().GetSupportedChains(ctx)[0]
+		index := chain.ChainName.String() + "_migration_tx_index"
+		k.GetObserverKeeper().SetFundMigrator(ctx, types.TssFundMigratorInfo{
+			ChainId:            chain.ChainId,
+			MigrationCctxIndex: index,
+		})
+		cctx := sample.CrossChainTx(t, index)
+		cctx.CctxStatus.Status = crosschaintypes.CctxStatus_OutboundMined
+		k.SetCrossChainTx(ctx, *cctx)
+
+		require.Equal(t, len(k.GetObserverKeeper().GetAllTssFundMigrators(ctx)), 1)
 		_, err := msgServer.UpdateTssAddress(ctx, &crosschaintypes.MsgUpdateTssAddress{
 			Creator:   admin,
 			TssPubkey: tssNew.TssPubkey,
 		})
-		assert.ErrorContains(t, err, "cannot update tss address not enough migrations have been created and completed")
-		assert.ErrorIs(t, err, crosschaintypes.ErrUnableToUpdateTss)
+		require.ErrorContains(t, err, "cannot update tss address not enough migrations have been created and completed")
+		require.ErrorIs(t, err, crosschaintypes.ErrUnableToUpdateTss)
 		tss, found := k.GetObserverKeeper().GetTSS(ctx)
-		assert.True(t, found)
-		assert.Equal(t, tssOld, tss)
+		require.True(t, found)
+		require.Equal(t, tssOld, tss)
 		migrators := k.GetObserverKeeper().GetAllTssFundMigrators(ctx)
 		assert.Equal(t, 1, len(migrators))
 	})
+
 	t.Run("unable to update tss when pending cctx is present", func(t *testing.T) {
 		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 		admin := sample.AccAddress()
@@ -146,9 +154,12 @@ func TestMsgServer_UpdateTssAddress(t *testing.T) {
 		msgServer := keeper.NewMsgServerImpl(*k)
 		tssOld := sample.Tss()
 		tssNew := sample.Tss()
+
 		k.GetObserverKeeper().SetTSSHistory(ctx, tssOld)
 		k.GetObserverKeeper().SetTSSHistory(ctx, tssNew)
 		k.GetObserverKeeper().SetTSS(ctx, tssOld)
+		setSupportedChain(ctx, zk, getValidEthChainIDWithIndex(t, 0), getValidEthChainIDWithIndex(t, 1))
+
 		for _, chain := range k.GetObserverKeeper().GetSupportedChains(ctx) {
 			index := chain.ChainName.String() + "_migration_tx_index"
 			k.GetObserverKeeper().SetFundMigrator(ctx, types.TssFundMigratorInfo{
@@ -172,6 +183,7 @@ func TestMsgServer_UpdateTssAddress(t *testing.T) {
 		migrators := k.GetObserverKeeper().GetAllTssFundMigrators(ctx)
 		assert.Equal(t, len(k.GetObserverKeeper().GetSupportedChains(ctx)), len(migrators))
 	})
+
 	t.Run("unable to update tss cctx is not present", func(t *testing.T) {
 		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
 		admin := sample.AccAddress()
@@ -179,9 +191,12 @@ func TestMsgServer_UpdateTssAddress(t *testing.T) {
 		msgServer := keeper.NewMsgServerImpl(*k)
 		tssOld := sample.Tss()
 		tssNew := sample.Tss()
+
 		k.GetObserverKeeper().SetTSSHistory(ctx, tssOld)
 		k.GetObserverKeeper().SetTSSHistory(ctx, tssNew)
 		k.GetObserverKeeper().SetTSS(ctx, tssOld)
+		setSupportedChain(ctx, zk, getValidEthChainIDWithIndex(t, 0), getValidEthChainIDWithIndex(t, 1))
+
 		for _, chain := range k.GetObserverKeeper().GetSupportedChains(ctx) {
 			index := chain.ChainName.String() + "_migration_tx_index"
 			k.GetObserverKeeper().SetFundMigrator(ctx, types.TssFundMigratorInfo{
